Combine test of buildmod to single file.
